Some import changes, made things more clear
