+++ Logic_layer/EmployeeLogic.py
@@ -1,9 +1,9 @@
-from Data_Layer.EmployeeData import EmployeeData
+from Data_Layer.DataWrapper import DataWrapper
 from Models.employee import Employee
 
 class EmployeeLogic:
-    def __init__(self, employee_data: EmployeeData):
-        self.employee_data = employee_data
+    def __init__(self):
+        self.data_wrapper = DataWrapper()
 
     def register_employee(self, employee_details: dict) -> str:
         new_employee = Employee(
@@ -16,18 +16,18 @@ class EmployeeLogic:
             location=employee_details["location"],
             is_manager=employee_details["is_manager"]
         )
-        self.employee_data.register_employee(new_employee)
+        self.data_wrapper.add_employee(new_employee)
         return "Employee registered successfully."
 
     def list_employees(self) -> list[Employee]:
-        return self.employee_data.get_all_employees()
+        return self.data_wrapper.list_employees()
 
     def change_employee_info(self, ssn: str, field: str, new_value: str) -> str:
-        self.employee_data.change_employee_info(ssn, field, new_value)
+        self.data_wrapper.update_employee(ssn, field, new_value)
         return "Employee information updated successfully."
 
     def search_employee_by_id(self, ssn: str) -> Employee:
-        employees = self.employee_data.get_all_employees()
+        employees = self.data_wrapper.list_employees()
         for employee in employees:
             if employee.ssn == ssn:
                 return employee
