+++ Logic_layer/PropertyLogic.py
@@ -1,9 +1,9 @@
-from Data_Layer.PropertyData import PropertyData
+from Data_Layer.DataWrapper import DataWrapper
 from Models.property import Property
 
 class PropertyLogic:
-    def __init__(self, property_data: PropertyData):
-        self.property_data = property_data
+    def __init__(self):
+        self.data_wrapper = DataWrapper()
 
     def add_property(self, property_details: dict) -> str:
         new_property = Property(
@@ -14,15 +14,15 @@ class PropertyLogic:
             manager=property_details["manager"],
             requires_maintenance=property_details["requires_maintenance"]
         )
-        self.property_data.add_property(new_property)
+        self.data_wrapper.add_property(new_property)
         return "Property added successfully."
 
     def list_properties(self) -> list[Property]:
-        return self.property_data.get_all_properties()
+        return self.data_wrapper.list_properties()
 
     def update_property(self, property_id, updated_details: dict) -> str:
-        self.property_data.update_property(property_id, updated_details)
+        self.data_wrapper.update_property(property_id, updated_details)
         return "Property updated successfully."
 
     def get_property_by_id(self, property_id):
-        return self.property_data.get_property_by_id(property_id)
+        return self.data_wrapper.get_property_by_id(property_id)
