+++ Logic_layer/WorkOrderLogic.py
@@ -3,8 +3,8 @@ from Models.WorkOrder import WorkOrder
 
 
 class WorkOrderLogic:
-    def __init__(self, data_wrapper: DataWrapper):
-        self.data_wrapper = data_wrapper
+    def __init__(self):
+        self.data_wrapper = DataWrapper()
 
     def CreateWorkOrder(self, work_order_obj: WorkOrder):
         """Takes in a WorkOrder object and forwards it to the data layer.
